pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saas-data-analytics-backend"
version = "0.1.0"
description = "Phase 1 backend foundation for research-oriented SaaS data analytics platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.0",
  "psycopg2-binary>=2.9.9",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt==4.0.1",
  "python-dotenv>=1.0.1",
  "pydantic-settings>=2.3.0",
  "email-validator>=2.2.0",
  "python-multipart>=0.0.9",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]

[tool.uv]
package = false
